requirements.txt
Flask==0.10.1
Flask-Cors==1.7.4
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
gunicorn==19.1.1
httplib2==0.9
itsdangerous==0.24
pymongo==2.7.2
python-instagram==1.1.3
simplejson==3.6.3
six==1.7.3
wsgiref==0.1.2
